pyproject.toml
[project]
name = "push-tmux"
version = "0.1.0"
description = "Pushbulletのメッセージをtmuxに送信するCLIツール"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "toml>=0.10.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "questionary>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-env>=1.0.0",
]

[project.scripts]
push-tmux = "push_tmux:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]
